lab6/product-service/requirements.txt
fastapi
uvicorn
motor
python-jose[cryptography]
passlib
python-multipart
aioredis
pymongo
pyjwt
confluent-kafka